Revert "[Enhancement] Integrate Backend-Based Pyodide Code Interpreter into Right Pane"
