+++ client/src/hooks/index.ts
@@ -18,7 +18,6 @@ export * from './ThemeContext';
 export * from './ScreenshotContext';
 export * from './ApiErrorBoundaryContext';
 export * from './Endpoint';
-export * from './usePyodide';
 
 export type { TranslationKeys } from './useLocalize';
