--- client/src/hooks/usePyodide/index.ts
@@ -1,1 +1,0 @@
-export { default } from './usePyodide';
