--- client/src/hooks/usePyodide/index.ts.bak
@@ -1,2 +1,0 @@
-// Export the usePyodide hook
-export { default as usePyodide } from './usePyodide';
